Move other data provider to correct folder
